fix(ui,web): the agents-list gear manages servers everywhere, not just desktop
On web the gear navigated to {key: 'settings', tab: 'agent-servers'} —
desktop's Settings window, a route the web app cannot open — so it
toasted "failed to open route" and there was no way to add your own
agent servers.

AgentServersSettings moves from desktop's settings page into
@shm/ui/agents/server-settings unchanged (it already used only shared
hooks), with the built-in default URL now coming from the platform seam
instead of the desktop-only constant. The gear goes through a new
optional seam member, useOpenServerSettings: desktop supplies its
settings-window navigation and behaves exactly as before; web omits it,
so the shared list opens the same management UI in a dialog right on
the agents page. The web adapter test now asserts the omission, and the
hover-only "Make default" reveal stays visible at phone widths.

Verified live in the browser: the gear opens the Agent Servers dialog,
adding a server toasts and the page's server list picks it up
immediately. Desktop 688/688, ui 314/314, web 161 passed, typecheck and
format clean.
